// models/raw-row.model.ts
export class RawRowModel {
    accountNumber: string = '';
    balance: string = '';
    cardNumber: string = '';
    cardType: string = '';
    cvv: string = '';
    expires: string = '';
    forename: string = '';
    hasChip: string = '';
    pan: string = '';
    sortCode: string = '';
    surname: string = '';
}